refactor: create component folder for LabVideoEdit page
