+++ YCAI/src/components/dashboard/lab/LabVideoEdit/VideoRecommendations.tsx
@@ -7,12 +7,12 @@ import * as QR from 'avenger/lib/QueryResult';
 import { declareQueries } from 'avenger/lib/react';
 import { pipe } from 'fp-ts/lib/function';
 
-import { updateRecommendationsForVideo } from '../../../state/dashboard/creator.commands';
-import * as queries from '../../../state/dashboard/public.queries';
+import { updateRecommendationsForVideo } from '../../../../state/dashboard/creator.commands';
+import * as queries from '../../../../state/dashboard/public.queries';
 import { ErrorBox } from '@shared/components/Error/ErrorBox';
-import { LazyFullSizeLoader } from '../../common/FullSizeLoader';
-import { RecommendationCard } from './RecommendationCard';
-import { ReorderList } from '../../common/ReorderList';
+import { LazyFullSizeLoader } from '../../../common/FullSizeLoader';
+import { RecommendationCard } from '../RecommendationCard';
+import { ReorderList } from '../../../common/ReorderList';
 
 const withQueries = declareQueries({
   settings: queries.settings,
