+++ YCAI/src/components/dashboard/lab/LabVideoEdit/index.tsx
@@ -1,8 +1,8 @@
 import { Grid } from '@material-ui/core';
 import * as React from 'react';
 import { addRecommendationForVideo } from 'state/dashboard/creator.commands';
-import { YTVideo } from '../../common/YTVideo';
-import AddRecommendationBox from './AddRecommendationBox';
+import { YTVideo } from '../../../common/YTVideo';
+import AddRecommendationBox from '../AddRecommendationBox';
 import { VideoRecommendations } from './VideoRecommendations';
 
 interface LabVideoEditProps {
